Explicitely use EnumSets and EnumAffix, Compute Preset incompats once and cache, rewrite NBT to use a BitSet, add config for Vortex to affect non-players

+++ src/main/java/c4/champions/common/affix/affix/AffixLively.java
@@ -44,7 +44,7 @@ public class AffixLively extends AffixBase {
         if (!entity.world.isRemote) {
             AffixNBT.Integer lastAttackTime = AffixNBT.getData(cap, this.getIdentifier(), AffixNBT.Integer.class);
 
-            if ((lastAttackTime.num + ConfigHandler.affix.lively.cooldown * 20) < entity.world.getTotalWorldTime()
+            if ((lastAttackTime.num + ConfigHandler.affix.lively.cooldown * 20) < (int)entity.world.getTotalWorldTime()
                     && entity.ticksExisted % 20 == 0) {
                 double healAmount = ConfigHandler.affix.lively.healAmount;
